requirements.txt
app==0.0.1
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.5
django-allauth==0.44.0
django-etc==1.3.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
filelock==3.0.12
Flask==2.0.1
idna==2.10
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.1
lark-parser==0.11.1
MarkupSafe==2.0.1
oauthlib==3.1.0
packaging==21.2
Pillow==8.1.0
playsound==1.2.2
psycopg2==2.8.6
psycopg2-binary==2.9.1
pycparser==2.20
pygame==2.0.0
PyJWT==2.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
required==0.4.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==2.0.1
